@@ -4,6 +4,7 @@
   "version": "0.0.0-dev",
   "author": "Ephigenia M. Eichner",
   "license": "MIT",
+  "type": "module",
   "homepage": "https://github.com/Ephigenia/notable-cli/",
   "keywords": [
     "cli",
@@ -25,8 +26,7 @@
     }
   ],
   "engines": {
-    "node": ">=14",
-    "npm": ">=6.14"
+    "node": ">=16.0.0"
   },
   "repository": {
     "type": "git",
@@ -51,11 +51,11 @@
     "lint": "eslint source",
     "postversion": "git push && git push --tags",
     "preversion": "npm test && npm run lint && npm run test:outdated",
-    "start": "node $npm_package_main",
+    "start": "node source/notable-cli.js",
     "tdd": "npm run test -- --watch",
     "test:outdated": "npm outdated",
     "test:security": "npm audit",
-    "test": "mocha --check-leaks --exit \"./{,!(node_modules)/**}/*.test.js\"",
+    "test": "mocha --check-leaks --exit \"./source/**/*.test.js\"",
     "update": "npx npm-check --update --save-exact",
     "version": "npm run changelog && git add CHANGELOG.md"
   },
@@ -65,14 +65,14 @@
     "dayjs": "1.10.7",
     "handlebars": "4.7.7",
     "marked": "4.0.10",
-    "marked-terminal": "4.2.0",
+    "marked-terminal": "5.0.0",
     "neo-blessed": "0.2.0",
-    "parse-md": "2.0.4",
+    "parse-md": "3.0.3",
     "recursive-readdir": "2.2.2"
   },
   "devDependencies": {
     "chai": "4.3.4",
-    "eslint": "8.0.0",
-    "mocha": "9.1.2"
+    "eslint": "8.6.0",
+    "mocha": "9.1.3"
   }
 }
